+++ src/main/java/project/web/data/domain/Hotel.java
@@ -18,7 +18,7 @@ public class Hotel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "h_num")
-    private Integer hNum; // 호텔식별번호 - room 과 외래키 ( 부모 )
+    private Long hNum; // 호텔식별번호 - room 과 외래키 ( 부모 )
 
     @Column(name = "h_name")
     private String hName; // 이름
@@ -46,7 +46,7 @@ public class Hotel {
     private String hAddr; // 주소
 
     @Column(name = "h_rate")
-    private Double hRate;
+    private Double hRate; // 별점
 
     @Column(name = "h_review")
     private Integer hReview;
